fix(tray): refresh calendar day labels
