+++ apps/desktop/src/sidebar/timeline/upcoming-meeting.ts
@@ -11,6 +11,7 @@ import {
 import { useIgnoredEvents } from "~/calendar/ignored-events";
 import { useTimelineTables } from "~/calendar/queries";
 import { useConfigValue } from "~/shared/config";
+import { useCurrentDay } from "~/shared/hooks/useCurrentDay";
 
 const UPCOMING_MEETING_VISIBLE_WINDOW_MS = 5 * 60 * 1000;
 const UPCOMING_MEETING_STATUS_TICK_MS = 1_000;
@@ -58,54 +59,6 @@ export function useSidebarUpcomingMeetingStatus({
   ]);
 
   return useUpcomingMeetingStatus(buckets, formatLabel, t`Now`);
-}
-
-function useCurrentDay(timezone: string | undefined) {
-  const store = useMemo(() => createCurrentDayStore(timezone), [timezone]);
-  return useSyncExternalStore(
-    store.subscribe,
-    store.getSnapshot,
-    store.getSnapshot,
-  );
-}
-
-function createCurrentDayStore(timezone: string | undefined) {
-  const getCurrentDay = () =>
-    new Intl.DateTimeFormat("en-CA", {
-      day: "2-digit",
-      month: "2-digit",
-      timeZone: timezone,
-      year: "numeric",
-    }).format(new Date());
-  let currentDay = getCurrentDay();
-
-  const subscribe = (listener: () => void) => {
-    const refresh = () => {
-      const nextDay = getCurrentDay();
-      if (nextDay === currentDay) return;
-      currentDay = nextDay;
-      listener();
-    };
-    const interval = setInterval(refresh, 60_000);
-    const handleVisibilityChange = () => {
-      if (document.visibilityState === "visible") refresh();
-    };
-
-    window.addEventListener("focus", refresh);
-    document.addEventListener("visibilitychange", handleVisibilityChange);
-    refresh();
-
-    return () => {
-      clearInterval(interval);
-      window.removeEventListener("focus", refresh);
-      document.removeEventListener("visibilitychange", handleVisibilityChange);
-    };
-  };
-
-  return {
-    getSnapshot: () => currentDay,
-    subscribe,
-  };
 }
 
 export function useUpcomingMeetingStatus(
